int bug fix

+++ sieve0.c
@@ -9,25 +9,26 @@
 #include "mpi.h"
 #include <math.h>
 #include <stdio.h>
+#include <stdlib.h>
 #define MIN(a,b)  ((a)<(b)?(a):(b))
 
 int main (int argc, char *argv[])
 {
-   long long int    count;        /* Local prime count */
+   unsigned long long int    count;        /* Local prime count */
    double elapsed_time; /* Parallel execution time */
-   long long int    first;        /* Index of first multiple */
-   long long int    global_count; /* Global prime count */
-   long long int    high_value;   /* Highest value on this proc */
-   int    i;
+   unsigned long long int    first;        /* Index of first multiple */
+   unsigned long long int    global_count; /* Global prime count */
+   unsigned long long int    high_value;   /* Highest value on this proc */
+   unsigned long long int    i;
    int    id;           /* Process ID number */
-   int    index;        /* Index of current prime */
-   long long int    low_value;    /* Lowest value on this proc */
+   unsigned long long int    index;        /* Index of current prime */
+   unsigned long long int    low_value;    /* Lowest value on this proc */
    char  *marked;       /* Portion of 2,...,'n' */
-   long long int    n;            /* Sieving from 2, ..., 'n' */
+   unsigned long long int    n;            /* Sieving from 2, ..., 'n' */
    int    p;            /* Number of processes */
-   long long int    proc0_size;   /* Size of proc 0's subarray */
-   int    prime;        /* Current prime */
-   int    size;         /* Elements in 'marked' */
+   unsigned long long int    proc0_size;   /* Size of proc 0's subarray */
+   unsigned long long int    prime;        /* Current prime */
+   unsigned long long int    size;         /* Elements in 'marked' */
 
    MPI_Init (&argc, &argv);
 
@@ -106,7 +107,7 @@ int main (int argc, char *argv[])
    /* Print the results */
 
    if (!id) {
-      printf ("There are %d primes less than or equal to %d\n",
+      printf ("There are %lld primes less than or equal to %lld\n",
          global_count, n);
       printf ("SIEVE (%d) %10.6f\n", p, elapsed_time);
    }
